Implemented ConfigParser and a simple Config object
ConfigParser is an utility object with functions to parse a configuration file
(JSON) into a Config instance using Argonaut. A simple config file has also
been added as a resource to be used as default configuration in a future.

Furthermore, Config object companion defines a type alias for the Reader monad
to be used with Config instances, and a utility method "withConfig" as
syntactic sugar hiding the usage of Reader. ClustalΩ functions have already
been updated in order to use that Configured and withConfig syntax.