Implement Fasta filter object
Filter object is responsible of filtering a set of Fasta files (from a
directory) based on a given similarity threshold. It scans all files, computes
a distance matrix in each one of them (using ClustalΩ) and search in the first
entry of that matrix if it has a value equal or higher than the given
threshold. If that is the case, the correspondent first entry is kept in the
filter, else it is discarded.

A thing to note is that it will only keep the first entry in each Fasta file.
The reason is because those files are supposed to be generated with the Fasta
Splitter object, by creating files with the first entry from a Fasta that wants
to be compared and the rest ones from a reference Fasta.